+++ zcash_client_backend/src/wallet.rs
@@ -30,9 +30,6 @@ use {
     std::time::SystemTime,
 };
 
-#[cfg(feature = "transparent-inputs")]
-pub mod transparent;
-
 /// A unique identifier for a shielded transaction output
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct NoteId {