+++ zcash_client_sqlite/src/error.rs
@@ -26,7 +26,9 @@ use crate::{
 use {
     crate::wallet::transparent::SchedulingError,
     ::transparent::{address::TransparentAddress, keys::TransparentKeyScope},
-    zcash_keys::encoding::TransparentCodecError,
+    zcash_keys::{
+        encoding::TransparentCodecError, keys::transparent::gap_limits::wallet::GapAddressesError,
+    },
 };
 
 /// The primary error type for the SQLite wallet backend.
@@ -166,6 +168,9 @@ pub enum SqliteClientError {
         actual: BlockHeight,
     },
 
+    #[cfg(feature = "transparent-inputs")]
+    GapAddresses,
+
     /// An attempt to import a transparent pubkey failed because that pubkey had already been
     /// imported to a different account.
     #[cfg(feature = "transparent-key-import")]
@@ -181,6 +186,17 @@ impl error::Error for SqliteClientError {
             SqliteClientError::BalanceError(e) => Some(e),
             SqliteClientError::AddressGeneration(e) => Some(e),
             _ => None,
+        }
+    }
+}
+
+#[cfg(feature = "transparent-inputs")]
+impl From<GapAddressesError<SqliteClientError>> for SqliteClientError {
+    fn from(err: GapAddressesError<SqliteClientError>) -> Self {
+        match err {
+            GapAddressesError::Storage(e) => e,
+            GapAddressesError::AddressGeneration(e) => SqliteClientError::AddressGeneration(e),
+            GapAddressesError::AccountUnknown => SqliteClientError::AccountUnknown,
         }
     }
 }
@@ -285,6 +301,13 @@ impl fmt::Display for SqliteClientError {
                     f,
                     "An address has already been exposed for diversifier index {}",
                     u128::from(*i)
+                )
+            }
+            #[cfg(feature = "transparent-inputs")]
+            SqliteClientError::GapAddresses => {
+                write!(
+                    f,
+                    "An error occured while generating a transparent gap addresses"
                 )
             }
             SqliteClientError::AddressReuse(address_str, txids) => {
@@ -397,6 +420,8 @@ impl From<PutBlocksError<SqliteClientError, commitment_tree::Error>> for SqliteC
             }
             ll::wallet::PutBlocksError::Storage(e) => e,
             ll::wallet::PutBlocksError::ShardTree(e) => SqliteClientError::from(e),
+            #[cfg(feature = "transparent-inputs")]
+            ll::wallet::PutBlocksError::GapAddresses(e) => SqliteClientError::from(e),
         }
     }
 }